Created a Proxy class to act as the base class for the FlashProxy and HTMLProxy classes to inherit from.
Fixed an issue with Flash autoplay that would not set the URL appropriately the first time play() was called externally. 

Added a FVideoModel.STATE_READY to be used to attach to the player once everything has initialized. This way the player is given a default starting state instead of nothing.